requirements.txt
entsoe-py
beautifulsoup4
requests
pandas
PyPDF2